Add files via upload
The goal of this assignment is to create a Discord bot that tracks a list of users and when they joined the server. The bot should track the list of users and join dates in a CSV file. Use the file A3_Start.py as a starting point.

When the bot starts for the first time it should create a list of users already registered and add them to a CSV with a date column that has the current date & time. Each time a user joins the server they should be added with their join date in the date column. Each time a user leaves the server, they should be removed from the CSV. The discord events on_member_join and on_member_remove should help with the last 2.

Create member list (5%)
Add new members to list (5%)
Remove member from list (5%)
Tips:

Create a second account to test adding/removing users on the server.
To get the current date and time in python use the built-in date time module like so

from datetime import datetime
print(datetime.now())


Remember date times need to be converted to strings before they can be saved to a CSV.
Use the file A3_Start.py to start your bot, it has some of the work done for you so you can focus on the important parts.
Don’t forget to turn on the Member Intent in the discord developer portal